feat: add courasel into date section and api for fetching cities
