+++ Components/layout.js
@@ -1,10 +1,12 @@
+import Footer from "./footer";
 import Navbar from "./navbar";
 
 export default function Layout({ children }) {
   return (
-    <div className=" flex flex-col h-screen">
+    <div className="flex flex-col">
       <Navbar />
       {children}
+      <Footer />
     </div>
   );
 }
