+++ package.json
@@ -21,18 +21,18 @@
     "dotenv": "^16.4.5",
     "flatpickr": "^4.6.13",
     "framer-motion": "^11.12.0",
-    "next": "^15.0.3",
-    "pg": "^8.13.1",
     "lucide": "^0.462.0",
+    "next": "^15.0.3",
     "next-themes": "^0.4.3",
+    "pg": "^8.13.1",
     "react": "^18.3.1",
     "react-datepicker": "^7.5.0",
     "react-dom": "^18.3.1",
     "react-flatpickr": "^3.10.13",
     "react-hook-form": "^7.53.2",
-    "webpack": "^5.96.1",
     "react-icons": "^5.3.0",
-    "swagger-ui-express": "^5.0.1"
+    "swagger-ui-express": "^5.0.1",
+    "webpack": "^5.96.1"
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
